third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/CustomElementRegistry-upgrade-expected.txt

This is a testharness.js-based test.
[FAIL] upgrade should upgrade a candidate element when called on a shadow root with an association
  assert_equals: expected "ABElement" but got "HTMLElement"
[FAIL] upgrade should not upgrade a candidate element not associated with a registry
  assert_equals: expected "ScopedABElement" but got "HTMLElement"
Harness: the test ran to completion.
